removing redundand namespace declarations
