+++ src/NServiceBus.Metrics/ProbeBuilders/CriticalTimeProbeBuilder.cs
@@ -1,32 +1,30 @@
-namespace NServiceBus.Metrics
+using System;
+using NServiceBus;
+using NServiceBus.Features;
+
+[ProbeProperties(ProbeType.Duration, "Critical Time", "Time between message sent till end of processing.")]
+class CriticalTimeProbeBuilder : DurationProbeBuilder
 {
-    using System;
-    using Features;
-
-    [ProbeProperties(ProbeType.Duration, "Critical Time", "Time between message sent till end of processing.")]
-    class CriticalTimeProbeBuilder : DurationProbeBuilder
+    public CriticalTimeProbeBuilder(FeatureConfigurationContext context)
     {
-        public CriticalTimeProbeBuilder(FeatureConfigurationContext context)
-        {
-            this.context = context;
-        }
-
-        protected override void WireUp(DurationProbe probe)
-        {
-            context.Pipeline.OnReceivePipelineCompleted(e =>
-            {
-                DateTime timeSent;
-                if (e.TryGetTimeSent(out timeSent))
-                {
-                    var endToEndTime = e.CompletedAt - timeSent;
-
-                    probe.Record(endToEndTime);
-                }
-
-                return CompletedTask;
-            });
-        }
-
-        readonly FeatureConfigurationContext context;
+        this.context = context;
     }
+
+    protected override void WireUp(DurationProbe probe)
+    {
+        context.Pipeline.OnReceivePipelineCompleted(e =>
+        {
+            DateTime timeSent;
+            if (e.TryGetTimeSent(out timeSent))
+            {
+                var endToEndTime = e.CompletedAt - timeSent;
+
+                probe.Record(endToEndTime);
+            }
+
+            return CompletedTask;
+        });
+    }
+
+    readonly FeatureConfigurationContext context;
 }
