+++ src/NServiceBus.Metrics/ProbeBuilders/DurationProbeBuilder.cs
@@ -1,39 +1,36 @@
-namespace NServiceBus.Metrics
+using System;
+using System.Reflection;
+using System.Threading.Tasks;
+
+/// <summary>
+/// Defines a custom metric.
+/// </summary>
+abstract class DurationProbeBuilder
 {
-    using System;
-    using System.Reflection;
-    using System.Threading.Tasks;
+    protected abstract void WireUp(DurationProbe probe);
 
-    /// <summary>
-    /// Defines a custom metric.
-    /// </summary>
-    abstract class DurationProbeBuilder
+    public DurationProbe Build()
     {
-        protected abstract void WireUp(DurationProbe probe);
+        var probe = GetProbe();
 
-        public DurationProbe Build()
-        {
-            var probe = GetProbe();
+        WireUp(probe);
 
-            WireUp(probe);
-
-            return probe;
-        }
-
-        DurationProbe GetProbe()
-        {
-            var attribute = GetType().GetCustomAttribute<ProbePropertiesAttribute>();
-
-            if(attribute == null)
-            {
-                var exceptionMessage = $"The type '{GetType()}' is not annotated with required '{typeof(ProbePropertiesAttribute).Name}'. This attribute has to be added to provide necessary metadata for the probe.";
-
-                throw new Exception(exceptionMessage);
-            }
-
-            return new DurationProbe(attribute.Name, attribute.Description);
-        }
-
-        protected static readonly Task<int> CompletedTask = Task.FromResult(0);
+        return probe;
     }
+
+    DurationProbe GetProbe()
+    {
+        var attribute = GetType().GetCustomAttribute<ProbePropertiesAttribute>();
+
+        if (attribute == null)
+        {
+            var exceptionMessage = $"The type '{GetType()}' is not annotated with required '{typeof(ProbePropertiesAttribute).Name}'. This attribute has to be added to provide necessary metadata for the probe.";
+
+            throw new Exception(exceptionMessage);
+        }
+
+        return new DurationProbe(attribute.Name, attribute.Description);
+    }
+
+    protected static readonly Task<int> CompletedTask = Task.FromResult(0);
 }
