+++ src/NServiceBus.Metrics/ProbeBuilders/ProbePropertiesAttribute.cs
@@ -1,25 +1,22 @@
-namespace NServiceBus.Metrics
+using System;
+
+[AttributeUsage(AttributeTargets.Class)]
+sealed class ProbePropertiesAttribute : Attribute
 {
-    using System;
-
-    [AttributeUsage(AttributeTargets.Class)]
-    sealed class ProbePropertiesAttribute : Attribute
+    public ProbePropertiesAttribute(ProbeType type, string name, string description)
     {
-        public ProbePropertiesAttribute(ProbeType type, string name, string description)
-        {
-            Type = type;
-            Name = name;
-            Description = description;
-        }
-
-        public readonly ProbeType Type;
-        public readonly string Name;
-        public readonly string Description;
+        Type = type;
+        Name = name;
+        Description = description;
     }
 
-    enum ProbeType
-    {
-        Signal,
-        Duration
-    }
+    public readonly ProbeType Type;
+    public readonly string Name;
+    public readonly string Description;
+}
+
+enum ProbeType
+{
+    Signal,
+    Duration
 }
