+++ src/NServiceBus.Metrics/Probes/DurationProbe.cs
@@ -1,23 +1,21 @@
-﻿namespace NServiceBus
+﻿using System;
+using NServiceBus;
+
+class DurationProbe : Probe, IDurationProbe
 {
-    using System;
-
-    class DurationProbe : Probe, IDurationProbe
+    public DurationProbe(string name, string description) : base(name, description)
     {
-        public DurationProbe(string name, string description) : base(name, description)
-        {
-        }
-
-        internal void Record(TimeSpan duration)
-        {
-            observers(duration);
-        }
-
-        public void Register(Action<TimeSpan> observer)
-        {
-            observers += observer;
-        }
-
-        Action<TimeSpan> observers = span => { };
     }
+
+    public void Register(Action<TimeSpan> observer)
+    {
+        observers += observer;
+    }
+
+    internal void Record(TimeSpan duration)
+    {
+        observers(duration);
+    }
+
+    Action<TimeSpan> observers = span => { };
 }
