+++ src/NServiceBus.Metrics/Probes/Probe.cs
@@ -1,0 +1,12 @@
+﻿abstract class Probe
+{
+    protected Probe(string name, string description)
+    {
+        Name = name;
+        Description = description;
+    }
+
+    public string Name { get; }
+
+    public string Description { get; }
+}
